correct time format

+++ srcs/src/u_helper.h
@@ -12,8 +12,8 @@
 void handle_sigint(int dummy);
 void u_setrunning(uint8_t mode, uint8_t * running);
 int u_help( void );
-int64_t u_timest(void);
-int64_t u_cmptime(long time);
+double u_timest(void);
+double u_cmptime(double time);
 int u_printpack(t_reply *pack, t_time * timer, char * ipstr, uint64_t seq);
 
 #endif
